+++ backend/driving-today/src/main/java/com/drivingtoday/domain/review/dto/ReviewRequest.java
@@ -1,4 +1,4 @@
-package com.drivingtoday.domain.review;
+package com.drivingtoday.domain.review.dto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
